+++ src/components/Calendar/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-10-08: Update its `border-color` and `border-radius` styles.
 - 2025-11-01: Make it use `utils/jss` and put its styles in CSS layer, etc.
 - 2025-11-26: Support the month/year picker.
+- 2025-11-28: Optimize month/year picker scrolling by using state-driven styles
+  and computing scroll distance from `event.deltaY`.